Move Param & Response definitions formatters under Endpoint namespace
